#32 ：DatabaseAccessor設計変更と関連コードの修正
`DatabaseAccessor` クラスが接続文字列ではなく `SQLiteConnection`
オブジェクトを受け取る設計に変更されました。これにより、接続の
管理が外部で行えるようになり、テストやリソース管理が容易になり
ました。テストコードもそれに合わせて修正され、インメモリデータ
ベースを使用する際に `SQLiteConnection` を直接渡す形に変更され
ています。

また、`StockInfo` クラスではメソッド名が `UpdateChartPrices`
から `SetupChartPrices` に変更され、可読性が向上しました。

さらに、プロジェクトファイルに `System.Data.SqlClient` パッケージ
が追加され、データベース操作の拡張が意図されています。
